fix bug #3
