+++ packages/client/src/api/forum/messages/models.ts
@@ -1,4 +1,4 @@
-export type TMessage = {
+export type TCreateMessage = {
   id?: number;
   text: string;
   authorId: number;
@@ -9,7 +9,16 @@ export type TMessage = {
   emojis: TEmoji[];
 };
 
-export type TMessageWithID = Omit<TMessage, 'id'> & { id: number };
+export type TMessage = {
+  id: number;
+  text: string;
+  authorId: number;
+  login: string;
+  postId: number;
+  date: Date;
+  mainMessageId: number;
+  emojis: TEmoji[];
+};
 
 export type TPostEmojiRequest = {
   postId: number;
