+++ packages/client/src/api/forum/posts/posts.ts
@@ -15,17 +15,20 @@ export const postsApi = createApi({
   baseQuery: fetchBaseQuery({
     ...backendAPISettings,
   }),
+  tagTypes: ['Posts'],
   endpoints: build => ({
     getPosts: build.query<TPost[], void>({
       query: () => ({ url: POSTS_ENDPOINTS.getPosts, method: 'GET' }),
+      providesTags: ['Posts'],
     }),
-    getPostById: build.mutation<TPost, number>({
+    getPostById: build.query<TPost, number>({
       query: id => ({ url: `${POSTS_ENDPOINTS.getPosts}/${id}`, method: 'GET' }),
     }),
     postPost: build.mutation<string, TPost>({
       query: data => ({ url: POSTS_ENDPOINTS.getPosts, method: 'POST', body: data }),
+      invalidatesTags: ['Posts'],
     }),
   }),
 });
 
-export const { useGetPostsQuery, useGetPostByIdMutation, usePostPostMutation } = postsApi;
+export const { useGetPostsQuery, useGetPostByIdQuery, usePostPostMutation } = postsApi;
