+++ packages/client/src/pages/Forum/Routes.tsx
@@ -5,11 +5,11 @@ import WithSuspense from 'src/hoc/WithSuspence';
 
 const { id, index } = paths.forum;
 const ForumPosts = WithSuspense(lazy(() => import('./pages/ForumPosts/ForumPosts')));
-const FormPost = WithSuspense(lazy(() => import('./pages/FormPost/FormPost')));
+const ForumPost = WithSuspense(lazy(() => import('./pages/ForumPost/ForumPost')));
 
 export const routes = [
   <Route index element={<ForumPosts />} key={index} />,
   <Route path={id} key={id}>
-    <Route index element={<FormPost />} />
+    <Route index element={<ForumPost />} />
   </Route>,
 ];
