--- packages/client/src/pages/Forum/pages/FormPost/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FormPost';
