+++ packages/client/src/pages/Forum/pages/ForumPost/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ForumPost';
