+++ packages/client/src/pages/Forum/pages/ForumPost/utils/emojiLinkIcon.ts
